+++ pages/_app.js
@@ -1,11 +1,15 @@
 import Layout from '../components/Layout'
 import '../styles/globals.css'
+import { Provider } from 'react-redux';
+import { store } from '../redux/store';
 
 function MyApp({ Component, pageProps }) {
   return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
+    <Provider store={store}>
+      <Layout>
+        <Component {...pageProps} />
+      </Layout>
+    </Provider>
 
   )
 }
