+++ redux/store.js
@@ -1,0 +1,12 @@
+// import {createStore, applyMiddleware} from 'redux'
+
+// const store = createStore() ;
+
+import { configureStore } from '@reduxjs/toolkit';
+import userReducer from './userSlice';
+
+export const store = configureStore({
+    reducer: {
+        user: userReducer
+    },
+});
